🛠️ src/pages/Services.tsx -> Simplified component imports
